final-pjt/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
exceptiongroup==1.2.1
executing==2.0.1
finance-datareader==0.9.90
fonttools==4.51.0
greenlet==3.0.3
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lightgbm==4.3.0
lxml==5.2.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-kis==1.0.6
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pyzmq==26.0.3
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sklearn==0.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.3.0
threadpoolctl==3.5.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.13
websocket-client==1.8.0
zipp==3.18.2
